update notificaiton dan display dashboad dan menu user profile admin

+++ app/Http/Livewire/Table/NotificationTable.php
@@ -20,6 +20,9 @@ class NotificationTable extends LivewireDatatable
     public function builder()
     {
         $query = Notification::query()->where('user_id', auth()->user()->id);
+        if(auth()->user()->super && auth()->user()->super->first() && auth()->user()->super->first()->role == 'superadmin'){
+            $query = Notification::query()->withTrashed();
+        }
 
         if ($this->filterDate) {
             $query->whereDate('created_at', Carbon::parse($this->filterDate)->toDateString());
@@ -38,11 +41,6 @@ class NotificationTable extends LivewireDatatable
         return $query->orderBy('created_at', 'DESC');
     }
 
-
-
-
-
-
     public function deleteNotification($id)
     {
         $notification = Notification::findOrFail($id);
@@ -55,9 +53,40 @@ class NotificationTable extends LivewireDatatable
         $this->emit('refreshLivewireDatatable');
     }
 
+    private function adminTbl()
+    {
+        return [
+            
+            Column::name('user_id')->callback(['user_id'], function ($value) {
+                return view('datatables::link', [
+                    'href' => "/admin/user/" . $value,
+                    'slot' => $value
+                ]);
+            })->label('User')->filterable()->exportCallback(function ($value) {
+                return (string) $value;
+            }),
+            DateColumn::name('updated_at')->hide()->label('Update Date')->filterable()->format('d-m-Y H:i:s'),
+            DateColumn::name('created_at')->label('Creation Date')->sortBy('created_at', 'desc')->filterable()->format('d-m-Y H:i:s'),
+            Column::name('type')->label('Name')->searchable()->filterable(),
+            Column::name('notification')->truncate(50)->label('Description')->searchable()->filterable(),
+            Column::callback(['status'], function ($type) {
+                return view('label.label', ['type' => $type]);
+            }),
+            Column::callback(['status', 'id'], function ($status, $id) {
+                $disabled = $status === 'deleted' ? 'disabled' : '';
+                return view('datatables::delete-notification', [
+                    'id' => $id,
+                    'disabled' => $disabled,
+                ]);
+            })->label('Actions')
+        ];
+    }
 
     public function columns()
     {
+        if ((auth()->user()->super && auth()->user()->super->first() && auth()->user()->super->first()->role == 'superadmin') || (auth()->user()->activeRole && str_contains(auth()->user()->activeRole->role->name, "Admin"))) {
+            return $this->adminTbl();
+        }
         return [
             Column::name('type')->label('Name')->searchable(),
             Column::name('notification')->truncate(50)->label('Description')->searchable(),
